split out installers
