+++ cli/src/installers/index.ts
@@ -1,10 +1,11 @@
 import { envVariablesInstaller } from "~/installers/envVars.js";
 import { nextAuthInstaller } from "~/installers/nextAuth.js";
 import { prismaInstaller } from "~/installers/prisma.js";
-import { strictEsLintAndPrettier } from "~/installers/strictEslintAndPrettier.js";
 import { tailwindInstaller } from "~/installers/tailwind.js";
 import { trpcInstaller } from "~/installers/trpc.js";
 import { type PackageManager } from "~/utils/getUserPkgManager.js";
+import { eslint } from "./eslint.js";
+import { prettier } from "./prettier.js";
 
 // Turning this into a const allows the list to be iterated over for programatically creating prompt options
 // Should increase extensability in the future
@@ -14,7 +15,8 @@ export const availablePackages = [
   "tailwind",
   "trpc",
   "envVariables",
-  "strictEslintAndPrettier",
+  "eslint",
+  "prettier",
 ] as const;
 export type AvailablePackages = (typeof availablePackages)[number];
 
@@ -58,8 +60,12 @@ export const buildPkgInstallerMap = (
     inUse: true,
     installer: envVariablesInstaller,
   },
-  strictEslintAndPrettier: {
-    inUse: packages.includes("strictEslintAndPrettier"),
-    installer: strictEsLintAndPrettier,
+  prettier: {
+    inUse: packages.includes("prettier"),
+    installer: prettier,
+  },
+  eslint: {
+    inUse: packages.includes("eslint"),
+    installer: eslint,
   },
 });
