+++ cli/src/installers/tailwind.ts
@@ -7,14 +7,7 @@ import { addPackageDependency } from "~/utils/addPackageDependency.js";
 export const tailwindInstaller: Installer = ({ projectDir, packages }) => {
   addPackageDependency({
     projectDir,
-    dependencies: [
-      "tailwindcss",
-      "postcss",
-      "autoprefixer",
-      "prettier",
-      "prettier-plugin-tailwindcss",
-      "@types/prettier",
-    ],
+    dependencies: ["tailwindcss", "postcss", "autoprefixer"],
     devMode: true,
   });
 
@@ -26,23 +19,12 @@ export const tailwindInstaller: Installer = ({ projectDir, packages }) => {
   const postcssCfgSrc = path.join(extrasDir, "config/postcss.config.cjs");
   const postcssCfgDest = path.join(projectDir, "postcss.config.cjs");
 
-  const prettierSrc = path.join(
-    extrasDir,
-    "config/prettier-with-tailwind.config.cjs",
-  );
-  const prettierDest = path.join(projectDir, "prettier.config.cjs");
-
   const cssSrc = path.join(extrasDir, "src/styles/globals.css");
   const cssDest = path.join(projectDir, "src/styles/globals.css");
 
   fs.copySync(twCfgSrc, twCfgDest);
   fs.copySync(postcssCfgSrc, postcssCfgDest);
   fs.copySync(cssSrc, cssDest);
-
-  // Let strictEslintAndPrettier install the prettier config
-  if (!packages?.strictEslintAndPrettier.inUse) {
-    fs.copySync(prettierSrc, prettierDest);
-  }
 
   // Remove vanilla css file
   const indexModuleCss = path.join(projectDir, "src/pages/index.module.css");
