--- cli/template/extras/config/strict-prettier-with-tailwind.config.cjs
@@ -1,9 +1,0 @@
-/** @type {import("prettier").Config} */
-const config = {
-  plugins: [
-    require.resolve("@ianvs/prettier-plugin-sort-import"),
-    require.resolve("prettier-plugin-tailwindcss"),
-  ],
-};
-
-module.exports = config;
